Create ElasticSearchQueryRoutes
