+++ delta/plugins/elasticsearch/src/main/scala/ch/epfl/bluebrain/nexus/delta/plugins/elasticsearch/routes/ElasticSearchViewsRoutes.scala
@@ -6,7 +6,6 @@ import akka.http.scaladsl.server._
 import ch.epfl.bluebrain.nexus.delta.plugins.elasticsearch.model.ElasticSearchViewRejection._
 import ch.epfl.bluebrain.nexus.delta.plugins.elasticsearch.model._
 import ch.epfl.bluebrain.nexus.delta.plugins.elasticsearch.model.permissions.{read => Read, write => Write}
-import ch.epfl.bluebrain.nexus.delta.plugins.elasticsearch.query.{DefaultSearchRequest, DefaultViewsQuery, ElasticSearchQueryError}
 import ch.epfl.bluebrain.nexus.delta.plugins.elasticsearch.{ElasticSearchViews, ElasticSearchViewsQuery}
 import ch.epfl.bluebrain.nexus.delta.rdf.Vocabulary
 import ch.epfl.bluebrain.nexus.delta.rdf.jsonld.context.JsonLdContext.keywords
@@ -20,24 +19,17 @@ import ch.epfl.bluebrain.nexus.delta.sdk.directives.DeltaDirectives._
 import ch.epfl.bluebrain.nexus.delta.sdk.directives.{AuthDirectives, DeltaSchemeDirectives}
 import ch.epfl.bluebrain.nexus.delta.sdk.fusion.FusionConfig
 import ch.epfl.bluebrain.nexus.delta.sdk.identities.Identities
-import ch.epfl.bluebrain.nexus.delta.sdk.identities.model.Caller
 import ch.epfl.bluebrain.nexus.delta.sdk.implicits._
 import ch.epfl.bluebrain.nexus.delta.sdk.marshalling.RdfMarshalling
 import ch.epfl.bluebrain.nexus.delta.sdk.model._
 import ch.epfl.bluebrain.nexus.delta.sdk.model.routes.Tag
-import ch.epfl.bluebrain.nexus.delta.sdk.model.search.AggregationResult.aggregationResultJsonLdEncoder
-import ch.epfl.bluebrain.nexus.delta.sdk.model.search.SearchResults.searchResultsJsonLdEncoder
-import ch.epfl.bluebrain.nexus.delta.sdk.model.search.{AggregationResult, PaginationConfig, SearchResults}
-import ch.epfl.bluebrain.nexus.delta.sdk.projects.FetchContext
 import ch.epfl.bluebrain.nexus.delta.sourcing.ProgressStatistics
-import ch.epfl.bluebrain.nexus.delta.sourcing.model.Label
 import ch.epfl.bluebrain.nexus.delta.sourcing.offset.Offset
 import ch.epfl.bluebrain.nexus.delta.sourcing.projections.Projections
 import io.circe.generic.semiauto.deriveEncoder
 import io.circe.syntax._
 import io.circe.{Encoder, Json, JsonObject}
 import kamon.instrumentation.akka.http.TracingDirectives.operationName
-import monix.bio.IO
 import monix.execution.Scheduler
 
 /**
@@ -53,8 +45,6 @@ import monix.execution.Scheduler
   *   the elasticsearch views query operations bundle
   * @param projections
   *   the projections module
-  * @param resourcesToSchemas
-  *   a collection of root resource segment with their corresponding schema
   * @param schemeDirectives
   *   directives related to orgs and projects
   * @param index
@@ -65,19 +55,15 @@ final class ElasticSearchViewsRoutes(
     aclCheck: AclCheck,
     views: ElasticSearchViews,
     viewsQuery: ElasticSearchViewsQuery,
-    defaultViewsQuery: DefaultViewsQuery.Elasticsearch,
     projections: Projections,
-    resourcesToSchemas: ResourceToSchemaMappings,
     schemeDirectives: DeltaSchemeDirectives,
     index: IndexingAction.Execute[ElasticSearchView]
 )(implicit
     baseUri: BaseUri,
-    paginationConfig: PaginationConfig,
     s: Scheduler,
     cr: RemoteContextResolution,
     ordering: JsonKeyOrdering,
-    fusionConfig: FusionConfig,
-    fetchContext: FetchContext[ElasticSearchQueryError]
+    fusionConfig: FusionConfig
 ) extends AuthDirectives(identities, aclCheck)
     with CirceUnmarshalling
     with ElasticSearchViewsDirectives
@@ -94,7 +80,7 @@ final class ElasticSearchViewsRoutes(
 
   def routes: Route =
     (baseUriPrefix(baseUri.prefix) & replaceUri("views", schema.iri)) {
-      concat(viewsRoutes, resourcesListings, aggregationResourceRoutes, genericResourcesRoutes)
+      viewsRoutes
     }
 
   private val viewsRoutes: Route =
@@ -279,138 +265,6 @@ final class ElasticSearchViewsRoutes(
       }
     }
 
-  private val genericResourcesRoutes: Route =
-    pathPrefix("resources") {
-      extractCaller { implicit caller =>
-        concat(
-          (searchParametersAndSortList & paginated) { (params, sort, page) =>
-            concat(
-              // List all resources
-              (pathEndOrSingleSlash & operationName(s"$prefixSegment/resources")) {
-                val request = DefaultSearchRequest.RootSearch(params, page, sort)
-                list(request)
-              },
-              (label & pathEndOrSingleSlash & operationName(s"$prefixSegment/resources")) { org =>
-                val request = DefaultSearchRequest.OrgSearch(org, params, page, sort)
-                list(request)
-              }
-            )
-          },
-          resolveProjectRef.apply { ref =>
-            projectContext(ref) { implicit pc =>
-              (searchParametersInProject & paginated) { (params, sort, page) =>
-                val request = DefaultSearchRequest.ProjectSearch(ref, params, page, sort)
-                concat(
-                  // List all resources inside a project
-                  (pathEndOrSingleSlash & operationName(s"$prefixSegment/resources/{org}/{project}")) {
-                    list(request)
-                  },
-                  idSegment { schema =>
-                    // List all resources inside a project filtering by its schema type
-                    (pathEndOrSingleSlash & operationName(s"$prefixSegment/resources/{org}/{project}/{schema}")) {
-                      underscoreToOption(schema) match {
-                        case None        => list(request)
-                        case Some(value) =>
-                          val r = DefaultSearchRequest.ProjectSearch(ref, params, page, sort, value)(fetchContext)
-                          list(r)
-                      }
-                    }
-                  }
-                )
-              }
-            }
-          }
-        )
-      }
-    }
-
-  private val aggregationResourceRoutes: Route =
-    pathPrefix("resources") {
-      extractCaller { implicit caller =>
-        (searchParametersAndSortList(baseUri) & paginated & aggregated) { (params, sort, page, _) =>
-          concat(
-            // Aggregate all resources
-            (pathEndOrSingleSlash & operationName(s"$prefixSegment/resources/aggregate")) {
-              val request = DefaultSearchRequest.RootSearch(params, page, sort)
-              aggregate(request)
-            },
-            // Aggregate all resources inside an organization
-            (label & pathEndOrSingleSlash & operationName(s"$prefixSegment/resources/{org}/aggregate")) { org =>
-              val request = DefaultSearchRequest.OrgSearch(org, params, page, sort)
-              aggregate(request)
-            },
-            resolveProjectRef.apply { ref =>
-              val request = DefaultSearchRequest.ProjectSearch(ref, params, page, sort)
-              concat(
-                // Aggregate all resources inside a project
-                (pathEndOrSingleSlash & operationName(s"$prefixSegment/resources/{org}/{project}/aggregate")) {
-                  aggregate(request)
-                }
-              )
-            }
-          )
-        }
-      }
-    }
-
-  private val resourcesListings: Route =
-    concat(resourcesToSchemas.value.map { case (Label(resourceSegment), resourceSchema) =>
-      pathPrefix(resourceSegment) {
-        extractCaller { implicit caller =>
-          concat(
-            (searchParametersAndSortList & paginated) { (params, sort, page) =>
-              concat(
-                // List all resources of type resourceSegment
-                (pathEndOrSingleSlash & operationName(s"$prefixSegment/$resourceSegment")) {
-                  val request = DefaultSearchRequest.RootSearch(params, page, sort, resourceSchema)(fetchContext)
-                  list(request)
-                },
-                // List all resources of type resourceSegment inside an organization
-                (label & pathEndOrSingleSlash & operationName(s"$prefixSegment/$resourceSegment/{org}")) { org =>
-                  val request = DefaultSearchRequest.OrgSearch(org, params, page, sort, resourceSchema)(fetchContext)
-                  list(request)
-                }
-              )
-            },
-            resolveProjectRef.apply { ref =>
-              projectContext(ref) { implicit pc =>
-                // List all resources of type resourceSegment inside a project
-                (searchParametersInProject & paginated & pathEndOrSingleSlash) { (params, sort, page) =>
-                  operationName(s"$prefixSegment/$resourceSegment/{org}/{project}") {
-                    val request =
-                      DefaultSearchRequest.ProjectSearch(ref, params, page, sort, resourceSchema)(fetchContext)
-                    list(request)
-                  }
-                }
-              }
-            }
-          )
-        }
-      }
-    }.toSeq: _*)
-
-  private def list(request: DefaultSearchRequest)(implicit caller: Caller): Route =
-    list(IO.pure(request))
-
-  private def list(request: IO[ElasticSearchQueryError, DefaultSearchRequest])(implicit caller: Caller): Route =
-    (get & paginated & extractUri) { (page, uri) =>
-      implicit val searchJsonLdEncoder: JsonLdEncoder[SearchResults[JsonObject]] =
-        searchResultsJsonLdEncoder(ContextValue(contexts.searchMetadata), page, uri)
-
-      emit(request.flatMap(defaultViewsQuery.list))
-    }
-
-  private def aggregate(request: DefaultSearchRequest)(implicit caller: Caller): Route =
-    aggregate(IO.pure(request))
-
-  private def aggregate(request: IO[ElasticSearchQueryError, DefaultSearchRequest])(implicit caller: Caller): Route =
-    get {
-      implicit val searchJsonLdEncoder: JsonLdEncoder[AggregationResult] =
-        aggregationResultJsonLdEncoder(ContextValue(contexts.aggregations))
-
-      emit(request.flatMap(defaultViewsQuery.aggregate))
-    }
-
   private val decodingFailedOrViewNotFound: PartialFunction[ElasticSearchViewRejection, Boolean] = {
     case _: DecodingFailed | _: ViewNotFound | _: InvalidJsonLdFormat => true
   }
@@ -427,28 +281,22 @@ object ElasticSearchViewsRoutes {
       aclCheck: AclCheck,
       views: ElasticSearchViews,
       viewsQuery: ElasticSearchViewsQuery,
-      defaultViewsQuery: DefaultViewsQuery.Elasticsearch,
       projections: Projections,
-      resourcesToSchemas: ResourceToSchemaMappings,
       schemeDirectives: DeltaSchemeDirectives,
       index: IndexingAction.Execute[ElasticSearchView]
   )(implicit
       baseUri: BaseUri,
-      paginationConfig: PaginationConfig,
       s: Scheduler,
       cr: RemoteContextResolution,
       ordering: JsonKeyOrdering,
-      fusionConfig: FusionConfig,
-      fetchContext: FetchContext[ElasticSearchQueryError]
+      fusionConfig: FusionConfig
   ): Route =
     new ElasticSearchViewsRoutes(
       identities,
       aclCheck,
       views,
       viewsQuery,
-      defaultViewsQuery,
       projections,
-      resourcesToSchemas,
       schemeDirectives,
       index
     ).routes
